필터 정리. 매니저 추가.

+++ Project/WinAPI/RandomFunction.h
@@ -6,7 +6,7 @@ class RandomFunction : public SingletonBase <RandomFunction>
 public:
 	RandomFunction()
 	{
-		srand(GetTickCount());
+		srand((unsigned int)GetTickCount());
 	}
 	~RandomFunction() {}
